// libs/finalization/ui/src/lib/adapters/queues-finalization-vm-props.interface.ts
import { TableRowBase } from '@cmz/shared-ui';

export interface QueuesFinalizationVmProps extends TableRowBase {
    uniqId: string;
    reportTypeLabel: string;
    operatorsLabel: string;
    sourceLabel: string;
    initiatorPhoneNumber: string;
    reportedAt: string;
    actionsRef: string;
}